Revert "Improved ui, fixed footer routing"
