+++ src/components/organisms/Home/Partnership.svelte
@@ -96,11 +96,6 @@
         return () => {
             clearInterval(logoAnimationInterval);
             clearInterval(logoAnimationResetInterval);
-            window.removeEventListener("focus", startAnimation);
-            window.removeEventListener("blur", () => {
-                clearInterval(logoAnimationInterval);
-                clearInterval(logoAnimationResetInterval);
-            });
         };
     });
 </script>
